consolidate _node_wrap tests
